requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
openai==0.28.0
python-dotenv==1.0.0
python-docx==0.8.11
gunicorn==20.1.0
Pillow==10.0.1
requests==2.31.0
huggingface_hub==0.16.4
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
urllib3==1.26.16
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
click==8.1.3
rq==1.13.0
redis==4.5.5
